Simplified the math and changed some other stuff
I noticed that (gunner_speed/(2*PI*r))*(2*PI) could be simplified by cancelling out 2*PI, so I did that. I also noticed that p_to_g_ang was consistently g_to_p_ang+PI so I changed that as well. That meant that the variables ndx and ndy were unnecessary so I took them out. I also changed so that the current phase is only printed out if the phase changes. For example, if the enemy is in Rush Phase for 5 seconds, instead of printing "Rush Phase" 50 times, it will only print it out once, because it never changed from Rush Phase.